Extract sidebar space content pager state
